[DEV-11989] Refactor integrations to adapters
