+++ packages/nextjs/src/adapters/analytics/server.ts
@@ -1,0 +1,17 @@
+import { type AsyncResolvable, resolveAsync } from '../../utils';
+
+export namespace AnalyticsAdapter {
+    export interface Configuration {
+        isTrackingEnabled: AsyncResolvable<boolean>;
+    }
+
+    export function connect(config: Configuration) {
+        async function useAnalytics() {
+            return {
+                isTrackingEnabled: resolveAsync(config.isTrackingEnabled),
+            };
+        }
+
+        return { useAnalytics };
+    }
+}
