+++ packages/nextjs/src/adapters/client.ts
@@ -1,0 +1,5 @@
+/* eslint-disable import/extensions */
+
+export * from './intl/client';
+export * from './routing/client';
+export * from './theme-settings/client';
