+++ packages/nextjs/src/adapters/environment/lib.ts
