+++ packages/nextjs/src/adapters/intl/lib/constants.ts
