+++ packages/nextjs/src/adapters/intl/lib/types.ts
