+++ packages/nextjs/src/adapters/intl/types.ts
