+++ packages/nextjs/src/adapters/metadata/lib/generateRootMetadata.ts
