+++ packages/nextjs/src/adapters/metadata/lib/index.ts
