+++ packages/nextjs/src/adapters/metadata/lib/types.ts
