+++ packages/nextjs/src/adapters/metadata/lib/utils/generateAlternateLanguageLinks.ts
