+++ packages/nextjs/src/adapters/metadata/lib/utils/index.ts
