+++ packages/nextjs/src/adapters/routing/client.ts
@@ -1,0 +1,3 @@
+export type * from './types';
+
+export { RoutingAdapter } from './lib/RoutingAdapter.client';
