+++ packages/nextjs/src/adapters/routing/lib/normalizeUrl.ts
