+++ packages/nextjs/src/adapters/routing/lib/types.ts
