+++ packages/nextjs/src/adapters/routing/server.ts
@@ -3,4 +3,4 @@ export type * from './types';
 export { createRouter as router } from './lib/createRouter.server';
 export { createRoute as route } from './lib/createRoute.server';
 
-export { integrateRouting } from './lib/integrateRouting.server';
+export { RoutingAdapter } from './lib/RoutingAdapter.server';
