+++ packages/nextjs/src/adapters/routing/types.ts
