+++ packages/nextjs/src/adapters/server.ts
@@ -1,0 +1,10 @@
+/* eslint-disable import/extensions */
+
+export * from './analytics/server';
+export * from './app-helper/server';
+export * from './environment/server';
+export * from './intl/server';
+export * from './metadata/server';
+export * from './prezly/server';
+export * from './routing/server';
+export * from './theme-settings/server';
