+++ packages/nextjs/src/adapters/types.ts
