--- packages/nextjs/src/integrations/client.ts
@@ -1,5 +1,0 @@
-/* eslint-disable import/extensions */
-
-export * from './intl/integrateIntl.client';
-export * from './routing/client';
-export * from './theme-settings/integrateThemeSettings.client';
