--- packages/nextjs/src/integrations/routing/client.ts
@@ -1,3 +1,0 @@
-export type * from './types';
-
-export { integrateRouting } from './lib/integrateRouting.client';
